use ConcurrentHashMap for UpdateAgent
This single data structure replaced the current two ConcurrentLinkedQueue, while maintaining the deterministic behavior of calling every entry once per iteration.
This is expected to resolve:
- the behavior of calling queue add/remove every update and set calls
- the memory overhead of ConcurrentLinkedQueue that it would create new Node objects every add/remove calls
The new ConcurrentHashMap maintains the current node of the entry on the update and set task, while it only modifies the node when adding/removing entry.